+++ kernel/src/main/scala/fun/airzihao/VectorSteam/VectorDistance/CosineSim.scala
@@ -8,9 +8,9 @@ package fun.airzihao.VectorSteam.VectorDistance
  */
 object CosineSim {
   // cosSim = (a*b)/(|a|*|b|)
-  def cosineSim(vector1: Array[Double], vector2: Array[Double]): Double = {
-    val numerator: Double = vector1.zip(vector2).map(pair => pair._1*pair._2).sum
-    val denominator: Double = EuclideanDis.eucLength(vector1) * EuclideanDis.eucLength(vector2)
+  def cosineSim(vector1: Array[Float], vector2: Array[Float]): Float = {
+    val numerator: Float = vector1.zip(vector2).map(pair => pair._1*pair._2).sum
+    val denominator: Float = EuclideanDis.eucLength(vector1) * EuclideanDis.eucLength(vector2)
     if (denominator == 0) 0
     else numerator / denominator
   }
