+++ kernel/src/main/scala/fun/airzihao/VectorSteam/VectorDistance/MinkowskiDis.scala
@@ -13,14 +13,14 @@ object MinkowskiDis {
 
 object ChebyshevDis {
   //  max( | x2-x1 | , | y2-y1 | )
-  def chebyDis(vector1: Array[Double], vector2: Array[Double]): Double = {
+  def chebyDis(vector1: Array[Float], vector2: Array[Float]): Float = {
     vector1.zip(vector2).map(pair => math.abs(pair._1 - pair._2)).max
   }
 }
 
 object ManhattanDis {
   //  sum( | x2-x1 | , | y2-y1 | )
-  def manhattanDis(vector1: Array[Double], vector2: Array[Double]): Double = {
+  def manhattanDis(vector1: Array[Float], vector2: Array[Float]): Float = {
     vector1.zip(vector2).map(pair => math.abs(pair._1 - pair._2)).sum
   }
 }
