+++ kernel/src/test/scala/fun/airzihao/VectorDistanceTest/JaccardSimTest.scala
@@ -10,12 +10,12 @@ import org.junit.{Assert, Test}
  * @Modified By:
  */
 class JaccardSimTest {
-  val v0: Array[Double] = Array(0.0.toDouble, 0)
-  val v1: Array[Double] = Array(1.0.toDouble, 1)
-  val v2: Array[Double] = Array(3.0.toDouble, 4.0.toDouble)
-  val v3: Array[Double] = Array(1.0.toDouble, 0, 0)
-  val v4: Array[Double] = Array(0, 1.0.toDouble, 0)
-  val v5: Array[Double] = Array(1.0.toDouble, 1, 0)
+  val v0: Array[Float] = Array(0.0.toFloat, 0)
+  val v1: Array[Float] = Array(1.0.toFloat, 1)
+  val v2: Array[Float] = Array(3.0.toFloat, 4.0.toFloat)
+  val v3: Array[Float] = Array(1.0.toFloat, 0, 0)
+  val v4: Array[Float] = Array(0, 1.0.toFloat, 0)
+  val v5: Array[Float] = Array(1.0.toFloat, 1, 0)
 
   @Test
   def test(): Unit = {
